libvirt driver: improve xen support for hvm and pv mode
This patch improves the libvirt.xml.template to generate a working
libvirt configuration for both hvm and pv instances. It sets correct
driver_names, which are required to making qcow2 images work for xen. It
also sets the os type based on the vm_mode and uses hvmloader for hvm
instances.

Change-Id: I5f92bedf6588ab2674d784772a5ecdb895fabe40